// src/server/domains/products/errors.types.ts
export enum ErrorKeys {
  PRODUCT_NOT_FOUND = 'PRODUCT_NOT_FOUND',
  PRODUCT_NAME_DUPLICATE = 'PRODUCT_NAME_DUPLICATE',
  INVALID_PRODUCT_DATA = 'INVALID_PRODUCT_DATA',
  MISSING_REQUIRED_FIELDS = 'MISSING_REQUIRED_FIELDS',
  INVALID_PACKAGE_TYPE = 'INVALID_PACKAGE_TYPE',
  INGREDIENT_NOT_FOUND = 'INGREDIENT_NOT_FOUND',
  BRAND_NOT_FOUND = 'BRAND_NOT_FOUND',
  UNIT_NOT_FOUND = 'UNIT_NOT_FOUND',
  FAILURE_TO_CREATE_PRODUCT = 'FAILURE_TO_CREATE_PRODUCT',
  FAILURE_TO_UPDATE_PRODUCT = 'FAILURE_TO_UPDATE_PRODUCT',
}